feat: wrapper로 pagination 마진 설정 등 JUL-36
