--- components/post/table/TableTemplate/TableTemplate.module.scss
@@ -1,23 +1,0 @@
-.tableContainer {
-  border: 1px solid $gray-20;
-  border-radius: 1rem;
-  font-size: 1.4rem;
-  color: $black;
-  overflow-x: auto;
-  overflow-y: overlay;
-  -webkit-overflow-scrolling: touch;
-
-  &::-webkit-scrollbar {
-    width: 0;
-    height: 0;
-  }
-  
-  &::-webkit-scrollbar-thumb,
-  &::-webkit-scrollbar-track {
-    background: transparent;
-  }
-  
-  @include mobile {
-    font-size: 1.2rem;
-  }
-}
