+++ components/post/table/index.ts
@@ -1,9 +1,9 @@
 import EmployeeTable from "./EmployeeTable/EmployeeTable";
 import EmployerTable from "./EmployerTable/EmployerTable";
-import TableTemplate from "./TableTemplate/TableTemplate";
+import Table from "./Table/Table";
 
 export {
   EmployeeTable,
   EmployerTable,
-  TableTemplate,
+  Table,
 };
